Use int64_t for size of mz_strm_mem

+++ mz_strm_mem.c
@@ -41,17 +41,17 @@ typedef struct mz_stream_mem_s {
     mz_stream stream;
     int32_t mode;
     uint8_t *buffer;   /* Memory buffer pointer */
-    int32_t size;      /* Size of the memory buffer */
-    int32_t limit;     /* Furthest we've written */
-    int32_t position;  /* Current position in the memory */
-    int32_t grow_size; /* Size to grow when full */
+    int64_t size;      /* Size of the memory buffer */
+    int64_t limit;     /* Furthest we've written */
+    int64_t position;  /* Current position in the memory */
+    int64_t grow_size; /* Size to grow when full */
 } mz_stream_mem;
 
 /***************************************************************************/
 
-static int32_t mz_stream_mem_set_size(void *stream, int32_t size) {
+static int32_t mz_stream_mem_set_size(void *stream, int64_t size) {
     mz_stream_mem *mem = (mz_stream_mem *)stream;
-    int32_t new_size = size;
+    int64_t new_size = size;
     uint8_t *new_buf = NULL;
 
     new_buf = (uint8_t *)malloc((uint32_t)new_size);
@@ -97,9 +97,9 @@ int32_t mz_stream_mem_read(void *stream, void *buf, int32_t size) {
     mz_stream_mem *mem = (mz_stream_mem *)stream;
 
     if (size > mem->size - mem->position)
-        size = mem->size - mem->position;
+        size = (int32_t)(mem->size - mem->position);
     if (mem->position + size > mem->limit)
-        size = mem->limit - mem->position;
+        size = (int32_t)(mem->limit - mem->position);
 
     if (size <= 0)
         return 0;
@@ -112,7 +112,7 @@ int32_t mz_stream_mem_read(void *stream, void *buf, int32_t size) {
 
 int32_t mz_stream_mem_write(void *stream, const void *buf, int32_t size) {
     mz_stream_mem *mem = (mz_stream_mem *)stream;
-    int32_t new_size = 0;
+    int64_t new_size = 0;
     int32_t err = MZ_OK;
 
     if (!size)
@@ -130,7 +130,7 @@ int32_t mz_stream_mem_write(void *stream, const void *buf, int32_t size) {
             if (err != MZ_OK)
                 return err;
         } else {
-            size = mem->size - mem->position;
+            size = (int32_t)(mem->size - mem->position);
         }
     }
 
@@ -199,7 +199,7 @@ int32_t mz_stream_mem_error(void *stream) {
     return MZ_OK;
 }
 
-void mz_stream_mem_set_buffer(void *stream, void *buf, int32_t size) {
+void mz_stream_mem_set_buffer(void *stream, void *buf, int64_t size) {
     mz_stream_mem *mem = (mz_stream_mem *)stream;
     mem->buffer = (uint8_t *)buf;
     mem->size = size;
@@ -223,17 +223,17 @@ int32_t mz_stream_mem_get_buffer_at_current(void *stream, const void **buf) {
     return mz_stream_mem_get_buffer_at(stream, mem->position, buf);
 }
 
-void mz_stream_mem_get_buffer_length(void *stream, int32_t *length) {
+void mz_stream_mem_get_buffer_length(void *stream, int64_t *length) {
     mz_stream_mem *mem = (mz_stream_mem *)stream;
     *length = mem->limit;
 }
 
-void mz_stream_mem_set_buffer_limit(void *stream, int32_t limit) {
+void mz_stream_mem_set_buffer_limit(void *stream, int64_t limit) {
     mz_stream_mem *mem = (mz_stream_mem *)stream;
     mem->limit = limit;
 }
 
-void mz_stream_mem_set_grow_size(void *stream, int32_t grow_size) {
+void mz_stream_mem_set_grow_size(void *stream, int64_t grow_size) {
     mz_stream_mem *mem = (mz_stream_mem *)stream;
     mem->grow_size = grow_size;
 }
